backend/requirements.txt
Django==5.2
django-crispy-forms==2.5
crispy-bootstrap5==2025.6

djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
python-decouple==3.8
pillow==12.0.0
pycountry==24.6.1
sorl-thumbnail==12.11.0

requests
stellar_sdk